Adding validation of servers during build steps
When jobs are saved, a hash of global config values is calculated, and saved
into the job's config.  At runtime, the saved value is compared against the
existing value.

Note:  The servername is used to find the corresponding cloud, and the
servername is selected from a list of all specific vsphere cloud objects.