ClusterSelection: Get datastores directly from cluster
In our Vsphere setup, searching for Datastores under the cluster doesn't work.
When calling getDataStoreClusters on a cluster, the cluster is found though.
Unfortunately I don't know enough about the vmware vim25 api to know why the
datastore wasn't found.
It is found if I don't pass in the cluster for the searching and the searching
takes  place on the root entity.
Pass in logger to Vsphere clone method so that output can be seen in jenkins
console.

While testing this I realized that several arguments are not expanded.
Cluster for instance is not expanded in either Clone or Deploy.
This would mean that ${CLUSTER} for instance wouldn't be expanded.
Improved logging messaging as well so errors in configuration are easier to find.