VsphereSelection: Allow dynamic selection of vsphere cloud
A jenkins user could have a job that can be used for two separate VSphere
instances.
At present there is no way to select the vsphere instance using a parameter.
This adds a parameter called ${VSPHERE_CLOUD_NAME} as an option for selection.
If selected then the jenkins user can add a parameter of the same name for the
build.