Disconnect from vSphere after slave is marked offline
In the earlier state, the plugin connected to vSphere without checking
if the slaveComputer was online and then looked at the power state of
the VM. If offline, it simply came out and forgot to disconnect (leaving
a session behind).

Wrapping the connection inside an if construct (checking if the slave
computer is online). Note that there might be a difference between a
slave computer state being anything but online and the VM still powered
on. This is a temporary solution to avoid huge number of requests
getting to the vSphere server.