VSphere folder parameter
A new folder parameter has been added for templates and clone/deploy
actions.

This parameter creates the new VM in the specified folder of vSphere
(which can be a path of folders or a unique name folder).

If left blank or folder not available, it will create the VM in the same
folder as the original VM.

It is specially interesting for organizing and separating templates from
autodeployed slaves.
For example, we have daily backups and we want separate folders for
templates and autodeployed dynamic slaves, so we can specify to backup
only the folders that contain the templates and omit folders that
contain the dynamic slaves.