Recognise VMs our Jenkins created and permit reconnection.
If we have previously lost contact with vSphere and thus end up with VMs
that are not known to Jenkins, but later we attempt to create a slave whose
same-named VM already exists, we can now recognise this situation and
reconnect.
Previously the code would repeatedly fail to create the VM and be unable to
spawn new VMs as a result.