vSphereCloudLauncher enhancement
vSphereCloudLauncher no longer connects to vSphere when its "idleAction" is
"nothing".  It still connects as normal if it needs to, it just doesn't bother
connecting when it's not going to use the connection.
It also now calls VSphere's disconnect method from a finally {...} block to
ensure that it will always disconnect from vSphere if it had previously
connected; before, it could remain connected if exceptions happened.