Retry deletion of unwanted VMs
Not every attempt to communicate with vSphere succeeds.  Sometimes we fail to
create a VM, sometimes we fail to destroy a VM.  If we fail to create one, the
existing logic will try again, as Jenkins will keep asking for new VMs.
If we failed to delete one, we used to merely log the failure exception and
then forget about it, which could result in unused VMs running in vSphere.

This code change adds extra in-memory state so that we remember a list of all
the VMs that we want to delete until we are sure that we have deleted them,
and we keep trying to delete unwanted VMs until we've got rid of them.