Fix VSphereCloudRetentionStrategy
The existing "keep until it's been idle for too long" strategy didn't appear
to be working - VMs would get disposed of immediately after becoming idle.
This change replaces the implementation with one that follows the techniques
used in the OpenStack plugin, as that one seems to work.
So, after this change, we now have a choice between a "use once" strategy or,
now this works, we can keep VMs around until they've been unused for too long
(which is configurable).
This allows us to get a better build throughput as we do not have to wait for
VMs to boot up for every build, as VMs remain available for reuse after each
build and only get killed off if they remain surplus to requirements for too
long.