Robustness improvements
Bugfix: vSphereCloudSlaveTemplate.provision method now ensures that it always disconnects from vSphere.  Previously there were some error cases where this might not happen.
Refactor: vSphereCloudSlaveTemplate.provision method now passes in the extra config parameters when the VM is initially created (instead of as a second vSphere operation).  This ensures that the meta-data (that we rely on for recognizing our own VMs) is there from the outset and hence we should now be less vulnerable to vSphere failing to power-on the VM after cloning it or to vSphere belatedly starting a VM it claimed to have failed to create.
Enhancement: If vSphereCloudSlaveTemplate.provision method failed to clone a new VM, we now attempt to delete the VM, as vSphere can partially-create a VM.
Enhancement: VSphere.cloneOrDeployVm method can (optionally) now set extra config parameters.